requirements.txt
asgiref==3.9.1
Django==4.2.23
djangorestframework==3.16.1
mysqlclient==2.2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
python-decouple==3.6
minio==7.1.14
Pillow==11.3.0
confluent-kafka==2.10.0
pylint==3.3.7
python-django==2.6.1